fixes
Merged memory and instructions into one list. When writing programs for UVSim, care should be taken that instructions aren't overwritten with STORE or READ commands.

Updated the unit tests to accommodate the changes to memory.

Changed the first READ command in test_invalid_instruction() to be a LOAD command so no input is requested.
Patched the input for test_successful_read() to have mock input from the test so that no input is requested at test runtime.

All unit tests should pass and no input should be required when running it for ease of testing.