Added external log writing
Added helper functions for writing to the output pane, writing to the external log file, and clearing the log file.
Everything that was being written to the output pane is now being written to an external log file called log.txt. This file is reset every time UVSimGUI4.py is launched.
I found that if you open log.txt in VSCode it will live update as you use the GUI, which is useful for debugging purposes.